reorganized and cleaned up the data importer code
