+++ frontend/src/metabase/visualizer/components/DataImporter/DataImporter.module.css
@@ -1,0 +1,25 @@
+.DataImporter {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-template-rows: 40px 1fr;
+  gap: var(--mantine-spacing-xs);
+  padding: var(--mantine-spacing-md);
+  grid-template-areas:
+    "title button"
+    "content content";
+  align-items: center;
+}
+
+.Title {
+  grid-area: title;
+  margin-bottom: 0 !important;
+}
+
+.ToggleButton {
+  grid-area: button;
+}
+
+.Content {
+  grid-area: content;
+  overflow: auto;
+}
