+++ frontend/src/metabase/visualizer/components/DataImporter/DataImporter.tsx
@@ -1,15 +1,34 @@
+import { useDisclosure } from "@mantine/hooks";
 import { useCallback, useState } from "react";
 import { t } from "ttag";
 
 import { useDebouncedValue } from "metabase/hooks/use-debounced-value";
 import { SEARCH_DEBOUNCE_DURATION } from "metabase/lib/constants";
-import { Box, Flex, Icon, TextInput } from "metabase/ui";
+import { useSelector } from "metabase/lib/redux";
+import {
+  Box,
+  Button,
+  Center,
+  Flex,
+  Icon,
+  Text,
+  TextInput,
+  Title,
+} from "metabase/ui";
+import { getDataSources } from "metabase/visualizer/selectors";
 
-import { SearchResultsList } from "./SearchResultsList";
+import { DatasetList } from "../DatasetList/DatasetList";
+
+import S from "./DataImporter.module.css";
+import { ResultsList } from "./ResultsList";
 
 export const DataImporter = () => {
   const [search, setSearch] = useState("");
   const [activeTab] = useState<string | null>("explore");
+  const [showDatasets, handlers] = useDisclosure(false);
+
+  const dataSources = useSelector(getDataSources);
+
   const debouncedSearch = useDebouncedValue(search, SEARCH_DEBOUNCE_DURATION);
 
   const handleSearchChange: React.ChangeEventHandler<HTMLInputElement> =
@@ -18,35 +37,78 @@ export const DataImporter = () => {
     }, []);
 
   return (
-    <Flex
-      direction="column"
-      style={{
-        height: "100%",
-      }}
+    <Box
+      className={S.DataImporter}
+      bg="var(--mb-color-bg)"
+      style={{ height: "100%" }}
+      data-testid="visualizer-data-importer"
     >
-      <Box>
-        <TextInput
-          m="xs"
-          variant="filled"
-          value={search}
-          onChange={handleSearchChange}
-          placeholder={t`Search for something`}
-          leftSection={<Icon name="search" />}
-        />
-      </Box>
-      <Flex
-        direction="column"
-        pt="sm"
-        px="sm"
-        style={{
-          overflowY: "auto",
-        }}
+      <Title order={4} mb="xs" className={S.Title}>
+        {showDatasets ? t`Add data` : t`Manage data`}
+      </Title>
+      <Button
+        size="xs"
+        variant="transparent"
+        ml="auto"
+        onClick={handlers.toggle}
+        className={S.ToggleButton}
+        aria-label={showDatasets ? t`Done` : t`Add more data`}
       >
-        <SearchResultsList
-          search={debouncedSearch}
-          mode={activeTab === "explore" ? "both" : "add"}
-        />
-      </Flex>
-    </Flex>
+        {showDatasets ? t`Done` : t`Add more data`}
+      </Button>
+
+      {showDatasets ? (
+        <Flex
+          direction="column"
+          className={S.Content}
+          style={{
+            height: "100%",
+          }}
+        >
+          <Box>
+            <TextInput
+              m="xs"
+              variant="filled"
+              value={search}
+              onChange={handleSearchChange}
+              placeholder={t`Search for something`}
+              leftSection={<Icon name="search" />}
+            />
+          </Box>
+          <Flex
+            direction="column"
+            pt="sm"
+            px="sm"
+            style={{
+              overflowY: "auto",
+            }}
+          >
+            <ResultsList
+              search={debouncedSearch}
+              mode={activeTab === "explore" ? "both" : "add"}
+            />
+          </Flex>
+        </Flex>
+      ) : (
+        <Flex
+          direction="column"
+          className={S.Content}
+          bg="white"
+          style={{
+            borderRadius: "var(--default-border-radius)",
+            height: "100%",
+            border: `1px solid var(--mb-color-border)`,
+          }}
+        >
+          {dataSources.length > 0 ? (
+            <DatasetList />
+          ) : (
+            <Center h="100%" w="100%" mx="auto">
+              <Text>{t`Pick a dataset first`}</Text>
+            </Center>
+          )}
+        </Flex>
+      )}
+    </Box>
   );
 };
