+++ frontend/src/metabase/visualizer/components/DataImporter/ResultsList.tsx
@@ -1,37 +1,115 @@
-import cx from "classnames";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useMemo } from "react";
 
-import { cardApi } from "metabase/api";
+import { skipToken, useListRecentsQuery, useSearchQuery } from "metabase/api";
+import { getDashboard } from "metabase/dashboard/selectors";
 import { useDispatch, useSelector } from "metabase/lib/redux";
-import { Box, Button, Flex, Icon, Text } from "metabase/ui";
-import { getVisualizerPrimaryColumn } from "metabase/visualizer/selectors";
-import { parseDataSourceId } from "metabase/visualizer/utils";
-import { isNumber, isString } from "metabase-lib/v1/types/utils/isa";
-import type { Field } from "metabase-types/api";
-import type {
-  VisualizerDataSource,
-  VisualizerDataSourceId,
-} from "metabase-types/store/visualizer";
+import { isNotNull } from "metabase/lib/types";
+import { Box, Loader } from "metabase/ui";
+import { getDataSources } from "metabase/visualizer/selectors";
+import { createDataSource } from "metabase/visualizer/utils";
+import {
+  addDataSource,
+  removeDataSource,
+} from "metabase/visualizer/visualizer.slice";
+import type { DashboardId, SearchResult } from "metabase-types/api";
+import type { VisualizerDataSource } from "metabase-types/store/visualizer";
 
-import S from "./ResultsList.module.css";
+import { Result } from "./Result";
 
-export interface ResultsListProps {
-  items: VisualizerDataSource[];
-  onSwap?: (item: VisualizerDataSource) => void;
-  onAdd?: (item: VisualizerDataSource) => void;
-  dataSourceIds: Set<VisualizerDataSourceId>;
+function shouldIncludeDashboardQuestion(
+  searchItem: SearchResult,
+  dashboardId: DashboardId | undefined,
+) {
+  return searchItem.dashboard ? searchItem.dashboard.id === dashboardId : true;
+}
+
+interface ResultsListProps {
+  search: string;
   mode: "swap" | "add" | "both";
 }
 
-export const ResultsList = ({
-  items,
-  onSwap,
-  onAdd,
-  dataSourceIds,
-  mode,
-}: ResultsListProps) => {
+export function ResultsList({ search, mode }: ResultsListProps) {
+  const dashboardId = useSelector(getDashboard)?.id;
+
+  const dispatch = useDispatch();
+  const dataSources = useSelector(getDataSources);
+  const dataSourceIds = useMemo(
+    () => new Set(dataSources.map(s => s.id)),
+    [dataSources],
+  );
+
+  const onAdd = useCallback(
+    (item: VisualizerDataSource) => {
+      if (dataSourceIds.has(item.id)) {
+        // remove data source if it exists
+        dispatch(removeDataSource(item));
+      } else {
+        // add data source
+        dispatch(addDataSource(item.id));
+      }
+    },
+    [dispatch, dataSourceIds],
+  );
+
+  const onSwap = useCallback(
+    (item: VisualizerDataSource) => {
+      // remove all data sources
+      dataSources.forEach(dataSource => {
+        dispatch(removeDataSource(dataSource));
+      });
+
+      // add data source
+      dispatch(addDataSource(item.id));
+    },
+    [dispatch, dataSources],
+  );
+
+  const { data: result = { data: [] } } = useSearchQuery(
+    search.length > 0
+      ? {
+          q: search,
+          limit: 10,
+          models: ["card"],
+          include_dashboard_questions: true,
+        }
+      : skipToken,
+    {
+      refetchOnMountOrArgChange: true,
+    },
+  );
+
+  const { data: allRecents = [] } = useListRecentsQuery(undefined, {
+    refetchOnMountOrArgChange: true,
+  });
+
+  const items = useMemo(() => {
+    if (
+      search.length === 0 ||
+      !Array.isArray(result.data) ||
+      result.data.length === 0
+    ) {
+      return allRecents
+        .filter(maybeCard =>
+          ["card", "dataset", "metric"].includes(maybeCard.model),
+        )
+        .map(card => createDataSource("card", card.id, card.name));
+    }
+    return result.data
+      .map(item =>
+        typeof item.id === "number" &&
+        shouldIncludeDashboardQuestion(item, dashboardId)
+          ? createDataSource("card", item.id, item.name)
+          : null,
+      )
+      .filter(isNotNull);
+  }, [result, allRecents, search, dashboardId]);
+
+  if (items.length === 0) {
+    return <Loader />;
+  }
+
   return (
-    <Box component="ul">
+    <Box component="ul" data-testid="results-list">
       {items.map((item, index) => (
         <Result
           key={index}
@@ -44,102 +122,4 @@ export const ResultsList = ({
       ))}
     </Box>
   );
-};
-
-interface ResultProps {
-  item: VisualizerDataSource;
-  onSwap?: (item: VisualizerDataSource) => void;
-  onAdd?: (item: VisualizerDataSource) => void;
-  selected: boolean;
-  mode: "swap" | "add" | "both";
 }
-
-const Result = (props: ResultProps) => {
-  const { selected, item, onSwap, onAdd, mode } = props;
-
-  const primaryColumn = useSelector(getVisualizerPrimaryColumn);
-
-  const [metadata, setMetadata] = useState<Field[] | undefined>([]);
-
-  const dispatch = useDispatch();
-
-  const getFieldsMetadata = useCallback(
-    async (id: VisualizerDataSourceId) => {
-      const { type, sourceId } = parseDataSourceId(id);
-      if (type === "card") {
-        const result = await dispatch(
-          cardApi.endpoints.getCard.initiate({ id: sourceId }),
-        );
-
-        setMetadata(result.data?.result_metadata);
-      }
-    },
-    [dispatch],
-  );
-
-  useEffect(() => {
-    getFieldsMetadata(item.id);
-  }, [item, getFieldsMetadata]);
-
-  const isCompatible = useMemo(() => {
-    if (!primaryColumn || !metadata) {
-      return true;
-    }
-
-    if (isNumber(primaryColumn) || isString(primaryColumn)) {
-      return metadata?.some(field => field.id === primaryColumn.id);
-    }
-
-    return metadata.some(field => field.base_type === primaryColumn?.base_type);
-  }, [metadata, primaryColumn]);
-
-  if (mode === "add" && !isCompatible) {
-    return null;
-  }
-
-  return (
-    <Flex
-      className={cx(S.resultItem, {
-        [S.resultItemSelected]: selected,
-      })}
-      align="center"
-      component="li"
-      px="sm"
-      py="sm"
-      mb="xs"
-      onClick={() => {
-        onSwap?.(item);
-      }}
-    >
-      <Icon
-        name="table2"
-        mr="xs"
-        style={{
-          flexShrink: 0,
-        }}
-      />
-      <Text size="md" truncate c="inherit" style={{ flexGrow: 1 }}>
-        {item.name}
-      </Text>
-      {mode === "both" && (
-        <div className={S.resultItemActions}>
-          <Icon name="refresh" className={S.refreshIcon} />
-
-          <Button
-            variant={selected ? "filled" : "inverse"}
-            size="xs"
-            rightSection={<Icon name="add" />}
-            onClick={e => {
-              e.stopPropagation();
-              onAdd?.(item);
-            }}
-            style={{
-              opacity: selected ? 0 : isCompatible ? 1 : 0.5,
-              pointerEvents: isCompatible ? "auto" : "none",
-            }}
-          />
-        </div>
-      )}
-    </Flex>
-  );
-};
