+++ frontend/src/metabase/visualizer/components/DatasetList/index.ts
@@ -1,2 +1,2 @@
 export * from "./ColumnListItem";
-export * from "./DataManager";
+export * from "./DatasetList";
